+++ packages/datatrak-web/src/features/CountrySelector/index.ts
@@ -4,4 +4,4 @@
  */
 
 export { useUserCountries } from './useUserCountries';
-export { CountrySelector } from './CountrySelector';
+export { CountrySelector, CountrySelectWrapper } from './CountrySelector';
