+++ packages/datatrak-web/src/views/index.ts
@@ -5,7 +5,7 @@
 
 export { LandingPage } from './LandingPage';
 export { SurveyPage } from './SurveyPage';
-export { SurveySelectPage } from './SurveySelectPage';
+export { SurveySelectPage } from './SurveySelectPage/SurveySelectPage.tsx';
 export { LoginPage } from './LoginPage';
 export { RegisterPage } from './RegisterPage';
 export { VerifyEmailResendPage } from './VerifyEmailResendPage';
